backend/python/requirements.txt
# Web scraping and data collection
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pytrends==4.9.2
lxml==4.9.3

# Google APIs
google-ads==22.1.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78
requests-oauthlib==1.3.1

# Data processing and storage
pandas==2.1.3
numpy==1.25.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Content generation
openai==1.3.7
markdown==3.5.1
jinja2==3.1.2

# API and web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
click==8.1.7
tqdm==4.66.1

# Development and testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0
